day18 微信支付功能 （1）下单完成，生成支付记录，调用微信API，获取支付链接，生成微信二维码 （2）生成二维码后， 循环检测用户是否支付，3秒一次，超过5min中，重新生成二维码，关闭页面不再生成 （3）检测支付完成后，修改订单状态，支付记录，清除支付记录缓存

+++ pinyougou_parent/pinyougou_pojo/src/main/java/com/pinyougou/pojo/TbPayLog.java
@@ -1,6 +1,7 @@
 package com.pinyougou.pojo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class TbPayLog implements Serializable {
@@ -10,7 +11,7 @@ public class TbPayLog implements Serializable {
 
     private Date payTime;
 
-    private Long totalFee;
+    private BigDecimal totalFee;
 
     private String userId;
 
@@ -46,11 +47,11 @@ public class TbPayLog implements Serializable {
         this.payTime = payTime;
     }
 
-    public Long getTotalFee() {
+    public BigDecimal getTotalFee() {
         return totalFee;
     }
 
-    public void setTotalFee(Long totalFee) {
+    public void setTotalFee(BigDecimal totalFee) {
         this.totalFee = totalFee;
     }
